requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
ruff==0.9.4
pytest==8.3.4
requests==2.32.3
pyjwt==2.10.1